+++ QuanLyDaiLy/Models/PhieuXuat.cs
@@ -7,7 +7,7 @@ public class PhieuXuat
     [Key]
     public int MaPhieuXuat { get; set; }
     public DateTime NgayLap { get; set; }
-    public double TongGiaTri { get; set; }
+    public long TongGiaTri { get; set; }
     //Foreign key
     public int MaDaiLy { get; set; }
     public virtual DaiLy DaiLy { get; set; } = null!;
